+++ packages/google_maps_flutter/google_maps_flutter/lib/google_maps_flutter.dart
@@ -29,7 +29,7 @@ export 'package:google_maps_flutter_platform_interface/google_maps_flutter_platf
         Heatmap,
         HeatmapGradient,
         HeatmapId,
-        GestureHandling,
+        WebGestureHandling,
         InfoWindow,
         JointType,
         LatLng,